pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zin-adk"
version = "0.1.3"
description = "ZAK — Zeron Agentic Kit, open-source ADK for building autonomous cybersecurity agents. Build, deploy, and govern autonomous cybersecurity agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "zeron.one", email = "info@zeron.one" }]
maintainers = [{ name = "zeron.one", email = "info@zeron.one" }]
keywords = [
    "cybersecurity",
    "security-agents",
    "ai-agents",
    "autonomous-agents",
    "risk-quantification",
    "devsecops",
    "soc-automation",
    "vulnerability-management",
    "compliance-automation",
    "security-intelligence",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "structlog>=23.0",
    "click>=8.0",
    "rich>=13.0",
    "python-ulid>=2.0",
]

[project.urls]
Homepage = "https://zeron.one"
Repository = "https://github.com/securezeron/zak"
"Bug Tracker" = "https://github.com/securezeron/zak/issues"
Changelog = "https://github.com/securezeron/zak/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "build>=1.0",
    "twine>=5.0",
]
# LLM providers — enables LLMAgent and reasoning.mode: llm_react
# Install with: pip install "zin-adk[llm]"
# For local/air-gapped use with Ollama, no extra packages are needed.
llm = [
    "openai>=1.30",                  # OpenAI GPT-4o / Azure OpenAI
    "anthropic>=0.28",               # Anthropic Claude
    "google-generativeai>=0.7",      # Google Gemini
]
# SIF graph backend — Memgraph via Bolt protocol (neo4j-python-driver).
# Run Memgraph: docker run -it -p 7687:7687 memgraph/memgraph
# Install with: pip install "zin-adk[graph]"
graph = [
    "neo4j>=5.0",
]
# Enterprise edition — set ZAK_EDITION=enterprise to unlock all 22 built-in agents.
# Platform (hosted enterprise) — PostgreSQL persistence + API key management.
# pip install "zin-adk[platform]"
# Required env: DATABASE_URL=postgresql+asyncpg://...  ZAK_EDITION=enterprise

[project.scripts]
zak = "zak.cli.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["zak*"]
exclude = ["zak.api*", "zak.platform*"]

[tool.setuptools.package-data]
"zak" = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
